dom/media/test/test_background_video_no_suspend_not_in_tree.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Test Background Video Doesn't Suspend When Timeout Is Longer Than Video</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<script src="manifest.js"></script>
<script src="background_video.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
<script>
"use strict";

var manager = new MediaTestManager;

var MIN_DELAY = 100;

/**
 * @param {string} url video src.
 * @returns {HTMLMediaElement} The created video element.
 */
function createVideoNotAppendToDoc(url, token, width, height) {
  // Default size of (160, 120) is used by other media tests.
  if (width === undefined) { width = 160; }
  if (height === undefined) { height = 3*width/4; }

  let v = document.createElement('video');
  v.token = token;
  v.width = width;
  v.height = height;
  v.src = url;
  return v;
}

startTest({
  desc: "Test Background Video Doesn't Suspend When If The Video Is Not In Tree.",
  prefs: [
    [ 'media.test.video-suspend', true ],
    [ 'media.suspend-background-video.enabled', true ],
    [ 'media.suspend-background-video.delay-ms', MIN_DELAY ]
  ],
  tests: gDecodeSuspendTests,
  runTest: (test, token) => {
    let v = createVideoNotAppendToDoc(test.name, token);
    manager.started(token);

    /* This test checks that suspend doesn't occur if a video element is not
       append to tree. */
    waitUntilPlaying(v)
      .then(() => checkVideoDoesntSuspend(v))
      .then(() => {
        ok(true, 'Video ended before decode was suspended');
        manager.finished(token); })
      .catch((e) => {
        ok(false, 'Test Failed: ' + e.toString());
        manager.finished(token); });
  }
});
</script>
